Re-factored create_dictionary.py to allow passing in the csv file from which to read the course data, and to append to a dictionary if one has already been created (by reading in the created file, and then adding the contents of the csv file to the dictionary). Note that if you append a file that has already been read, it will add the enrollment again; this should be fixed but right now we should make sure never to run the same file through twice.
Added code to make create_dictionary.py idempotent if run on the same csv file with a dictionary that has already added
the contents of that csv. When an existing dictionary is read, a set is created that will keep track of the course-id,
years, and term for the courses in the dictionary. If this combination is found in the csv being read in, it will be
skipped.

Wrote create_dictionary_new.py, which will convert the new form of .csv received from the SIS group to a dictionary that
is in the same form as that produced by create_dictionary.py. If I wanted to be clever, I could merge the code into
create dictionary (and add a flag or a configuration to determine which of the formats the original csv is in), but
since the format of the dumps are likely to change, I dicided just to write a new program. It does, however, re-use
much of the original program.

In doing this, I found a somewhat simpler way of determining if the dictionary entries have been created, and so went
back and changed create_dictionary.py to use the simpler algorithm.